requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==21.0
PyJWT==2.3.0
